requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-encrypted-model-fields==0.6.5
django-extensions==4.1
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.1
Markdown==3.10
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
